api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pymongo==4.5.0
python-dotenv==1.0.0